Technical Test
This application consists in two UIViewController. The first one has an
UITableViewController linked to CoreData. This table shows what is
stored order by rating. Each cell displays one picture, name and rating.
The second view is a detail visualization of each picture. It shows
name of the photo, description (maximum 3 lines), user avatar and user
name, name of the camera, aperture, lens, focus, iso and location. This
view only shows the available info.

Whenever the application is launched, it shows data that is currently
stored in the local SQLite and checks for an update automatically. If
there are a response, the application erases all content and stores the
new one. If the user scrolls at the bottom the the table it will ask
for more pictures.